+++ hs-bindgen/test-artefacts/fixtures/macros/redeclaration/same_line_tag_field/Example.hs
@@ -1,0 +1,127 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE MagicHash #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE StandaloneDeriving #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UnboxedTuples #-}
+{-# LANGUAGE UndecidableInstances #-}
+
+module Example
+    ( Example.B(..)
+    , Example.S(..)
+    )
+  where
+
+import qualified HsBindgen.Runtime.HasCField as HasCField
+import qualified HsBindgen.Runtime.Internal.Prelude as RIP
+import qualified HsBindgen.Runtime.Internal.Prelude.CompatHasField as RIP.CompatHasField
+import qualified HsBindgen.Runtime.Marshal as Marshal
+
+{-| __C declaration:__ @macro B@
+
+    __defined at:__ @macros\/redeclaration\/same_line_tag_field.h 14:9@
+
+    __exported by:__ @macros\/redeclaration\/same_line_tag_field.h@
+-}
+newtype B = B
+  { unwrapB :: RIP.CInt
+  }
+  deriving stock (Eq, RIP.Generic, Ord, Read, Show)
+  deriving newtype
+    ( RIP.Bitfield
+    , RIP.Bits
+    , Bounded
+    , Enum
+    , RIP.FiniteBits
+    , RIP.HasFFIType
+    , Integral
+    , RIP.Ix
+    , Num
+    , RIP.Prim
+    , Marshal.ReadRaw
+    , Real
+    , Marshal.StaticSize
+    , RIP.Storable
+    , Marshal.WriteRaw
+    )
+
+instance (ty ~ RIP.CInt) => RIP.CompatHasField.HasField "unwrapB" B ty where
+
+  hasField =
+    \x0 ->
+      (\y1 -> B {unwrapB = y1}, RIP.getField @"unwrapB" x0)
+
+instance ( ty ~ RIP.CInt
+         ) => RIP.HasField "unwrapB" (RIP.Ptr B) (RIP.Ptr ty) where
+
+  getField = HasCField.fromPtr (RIP.Proxy @"unwrapB")
+
+instance HasCField.HasCField B "unwrapB" where
+
+  type CFieldType B "unwrapB" = RIP.CInt
+
+  offset# = \_ -> \_ -> 0
+
+{-| __C declaration:__ @struct S@
+
+    __defined at:__ @macros\/redeclaration\/same_line_tag_field.h 15:8@
+
+    __exported by:__ @macros\/redeclaration\/same_line_tag_field.h@
+-}
+data S = S
+  { s_x :: B
+    {- ^ __C declaration:__ @x@
+
+         __defined at:__ @macros\/redeclaration\/same_line_tag_field.h 15:13@
+
+         __exported by:__ @macros\/redeclaration\/same_line_tag_field.h@
+    -}
+  }
+  deriving stock (Eq, RIP.Generic, Show)
+
+instance Marshal.StaticSize S where
+
+  staticSizeOf = \_ -> (4 :: Int)
+
+  staticAlignment = \_ -> (4 :: Int)
+
+instance Marshal.ReadRaw S where
+
+  readRaw =
+    \ptr0 ->
+          pure S
+      <*> HasCField.readRaw (RIP.Proxy @"s_x") ptr0
+
+instance Marshal.WriteRaw S where
+
+  writeRaw =
+    \ptr0 ->
+      \s1 ->
+        case s1 of
+          S s_x2 ->
+            HasCField.writeRaw (RIP.Proxy @"s_x") ptr0 s_x2
+
+deriving via Marshal.EquivStorable S instance RIP.Storable S
+
+instance (ty ~ B) => RIP.CompatHasField.HasField "s_x" S ty where
+
+  hasField =
+    \x0 -> (\y1 -> S {s_x = y1}, RIP.getField @"s_x" x0)
+
+instance (ty ~ B) => RIP.HasField "s_x" (RIP.Ptr S) (RIP.Ptr ty) where
+
+  getField = HasCField.fromPtr (RIP.Proxy @"s_x")
+
+instance HasCField.HasCField S "s_x" where
+
+  type CFieldType S "s_x" = B
+
+  offset# = \_ -> \_ -> 0